+++ src/route/clients/households/households_certificate_file_controller.ts
@@ -79,22 +79,27 @@ export async function updateCertificateFile(
   const payloadFile = payload.certificateFile;
   const payloadFileUrl = payloadFile.fileUrl;
 
-  const fileId = payloadFileUrl.split(DRIVE_URL_FILE_PATH)[1];
+  let fileId = null;
+  if (payloadFileUrl !== "") {
+    fileId = payloadFileUrl.split(DRIVE_URL_FILE_PATH)[1];
+  }
 
   if (
     payloadFile.mimeType != undefined &&
     payloadFile.rawData != undefined
   ) {
-    await deleteFile(
-      driveService,
-      fileId,
-    );
+    if (fileId !== null) {
+      await deleteFile(
+        driveService,
+        fileId,
+      );
+    }
 
     fileUrl = await uploadCertificateFile(
       payload,
       folderId
     );
-  } else {
+  } else if (fileId !== null) {
     const currentFile = await getFile(
       driveService,
       fileId,
